+++ clintools/urls.py
@@ -10,6 +10,7 @@ urlpatterns = [
     # url(r'^$', 'clintools.views.home', name='home'),
     # url(r'^blog/', include('blog.urls')),
     url(r'^pttrack/', include('pttrack.urls')),
+    url(r'^followup/', include('followup.urls')),
     url(r'^admin/', include(admin.site.urls)),
     url(r'^accounts/', include('django.contrib.auth.urls')),
     url(r'^$', RedirectView.as_view(pattern_name="home", permanent=False)),
